major imporvements but terrains are now cursed
