--- examples/classroom/authentication.html
@@ -1,37 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <link href="//maxcdn.bootstrapcdn.com/bootstrap/4.1.1/css/bootstrap.min.css" rel="stylesheet" id="bootstrap-css">
-    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.5.0/css/all.css">
-    <link rel="stylesheet" href="css/authentication.css">
-</head>
-<body>
-    <div class="container">
-        <div class="card card-login mx-auto text-center bg-dark">
-            <div class="mx-auto bg-dark">
-                <span id="logo"> 
-                    <img src="assets/logo.png" class="w-50" alt="Logo"> 
-                </span>
-                <br/>
-                <span class="card-header logo_title mt-5"> Authentication </span>
-                <br/>
-                <label id="error"></label>
-            </div>
-            <br/>
-            <div>
-                <button class="wrapper w-75" id="connect-wallet">
-                    Connect Wallet
-                </button>
-            </div>
-            <br/>
-            <!-- <div>
-                <button class="wrapper w-75" id="no-wallet">
-                    Without Wallet
-                </button>
-            </div> -->
-        </div>
-        <script src="js/authentication.js"></script>
-    </div>
-</body>
-
-</html>
