--- examples/classroom/css/app.css
@@ -1,4 +1,0 @@
-body {
-    margin: 0;
-    font-family: 'Space Mono', monospace;
-}
