+++ examples/dome/js/main.js
@@ -6,8 +6,8 @@ let virtualEnvironment = new VirtualEnvironment();
 // then fill your world with the stuff you want
 init();
 function init() {
-    virtualEnvironment.loadTerrain('../../../resources/terrains/dome.fbx', 0, 0, 0, "fbx", 0.2)
-    virtualEnvironment.spawnPlayer({spawn: {x:0, y:70, z:0}})
+    virtualEnvironment.loadTerrain('../../../resources/terrains/dome.fbx', 0, 0, 0, "fbx", 0.008)
+    virtualEnvironment.spawnPlayer({spawn: {x:0, y:10, z:0}})
 }
 
 // then start the animation
