+++ examples/floor-is-lava/js/main.js
@@ -8,15 +8,15 @@ let virtualEnvironment = new VirtualEnvironment();
 // then fill your world with the stuff you want
 init();
 function init() {
-    virtualEnvironment.spawnPlayer({spawn: { x:0, y:50, z:0}})
+    virtualEnvironment.spawnPlayer({spawn: { x:0, y:10, z:0}})
     
     setTimeout(() => {
-        virtualEnvironment.newSolidGeometriesFromSource('../../../resources/objects/rock_platform.glb', 0, 20, 1000, 16);
-        virtualEnvironment.newSolidGeometriesFromSource('../../../resources/objects/rock_platform.glb', 0, -20, -30, 16);
+        virtualEnvironment.newSolidGeometriesFromSource('../../../resources/objects/rock_platform.glb', 0, 0, 50, 2);
+        virtualEnvironment.newSolidGeometriesFromSource('../../../resources/objects/rock_platform.glb', 0, 0, 0, 2);
 
         for(let i=0; i<35; i++) {
-            let position = new THREE.Vector3(Math.random() * 800 - 400, 0, Math.random() * 1000);
-            virtualEnvironment.newSolidGeometriesFromSource('../../../resources/objects/rock_platform.glb', position.x, position.y, position.z, Math.random() * 20 + 4);
+            let position = new THREE.Vector3(Math.random() * 80 - 40, 0, Math.random() * 50);
+            virtualEnvironment.newSolidGeometriesFromSource('../../../resources/objects/rock_platform.glb', position.x, position.y, position.z, Math.random() * 3 + 1);
         }
 
         virtualEnvironment.terrainController.generateCollider(virtualEnvironment.scene);
