+++ index.html
@@ -7,13 +7,15 @@
         <h1>Welcome to the std kit</h1>
         <br>
         <p>here is a list of example implementations of the standard kit</p>
+        <h3>Examples</h3>
         <ol>
-            <li><a href="./examples/classroom/authentication.html">Classroom</a></li>
             <li><a href="./examples/base-template/index.html">Base Template</a></li>
             <li><a href="./examples/dome/index.html">Dome</a></li>
             <li><a href="./examples/floor-is-lava/index.html">Floor is Lava</a></li>
-            <li><a href="./examples/event-planet/index.html">parkour</a></li>
-            <li>Meta</li>
+        </ol>
+        <h3>Real usages</h3>
+        <ol>
+            <li><a href="./metaverse/event-planet/index.html">parkour</a></li>
             <li><a href="./metaverse/planet/index.html">planets</a></li>
             <li><a href="./metaverse/spawn-planet/index.html">Spawn Planet</a></li>
         </ol>
