+++ metaverse/event-planet/css/app.css
@@ -1,0 +1,4 @@
+body {
+    margin: 0;
+    font-family: 'Space Mono', monospace;
+}
